Refactor CND backend modules into package
