+++ backend/cnds/__init__.py
@@ -1,0 +1,4 @@
+"""Convenience exports for the CNDs package."""
+from .routes import router
+
+__all__ = ["router"]
