+++ backend/cnds/cnds_worker_anapolis.py
@@ -17,9 +17,11 @@ from playwright.async_api import async_playwright
 # Garante que variáveis do .env estejam disponíveis antes de qualquer uso
 _DOTENV_PATH = find_dotenv(filename=".env", raise_error_if_not_found=False)
 if not _DOTENV_PATH:
-    candidate = Path(__file__).resolve().parent / ".env"
-    if candidate.exists():
-        _DOTENV_PATH = str(candidate)
+    base_dir = Path(__file__).resolve().parent
+    for candidate in (base_dir / ".env", base_dir.parent / ".env"):
+        if candidate.exists():
+            _DOTENV_PATH = str(candidate)
+            break
 load_dotenv(dotenv_path=_DOTENV_PATH or None)
 
 # CORREÇÃO: Configurar event loop ANTES de qualquer import do FastAPI
